clean up unused files

+++ app/routes/api/forms/deletelike.tsx
@@ -1,22 +1,30 @@
 import type { ActionFunction } from "@remix-run/node";
-import { redirect } from "@remix-run/node";
 import invariant from "tiny-invariant";
 import { prisma } from "~/db.server";
+import { requireUser } from "~/session.server";
 
-export const action: ActionFunction = async (RemixContext) => {
-  const redirectURL = RemixContext.request.headers.get("Referer");
-  const formData = await RemixContext.request.formData();
-  const userId = formData.get("userId") as string;
-  const postId = formData.get("postId") as string;
+export const action: ActionFunction = async ({ request }) => {
+  try {
+    const redirectURL = request.headers.get("Referer");
+    const formData = await request.formData();
+    const { id } = await requireUser(request);
+    const userId = formData.get("userId") as string;
+    const postId = formData.get("postId") as string;
 
-  invariant(redirectURL, "Must have a original url");
-  await prisma.like.delete({
-    where: {
-      userId_postId: {
-        userId,
-        postId,
+    if (id !== userId) {
+      throw Error("User denied!");
+    }
+    invariant(redirectURL, "Must have a original url");
+    await prisma.like.delete({
+      where: {
+        userId_postId: {
+          userId,
+          postId,
+        },
       },
-    },
-  });
-  return { ok: true, message: "successfully deleted like" };
+    });
+    return { ok: true, message: "successfully deleted like" };
+  } catch (error: any) {
+    return { ok: false, message: error.message };
+  }
 };
